requirements.txt
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
xgboost>=2.0
pybaseball>=2.2.7
MLB-StatsAPI>=1.7.2
python-dotenv>=1.0.0
discord.py>=2.3
selenium>=4.20
requests>=2.31
joblib>=1.3
tqdm>=4.66
google-genai>=0.6